airflow_local/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
sqlalchemy
pandas-gbq
